Move extensions files into helpers package
