+++ android/app/src/main/java/io/freetubeapp/freetube/javascript/FreeTubeJavaScriptInterface.kt
@@ -17,30 +17,26 @@ import android.media.session.PlaybackState.STATE_PAUSED
 import android.net.Uri
 import android.os.Build
 import android.os.Build.VERSION.SDK_INT
-import android.os.Environment
 import android.provider.OpenableColumns
 import android.provider.Settings
 import android.view.WindowManager
 import android.webkit.JavascriptInterface
-import android.webkit.PermissionRequest
 import androidx.activity.result.ActivityResult
 import androidx.annotation.RequiresApi
 import androidx.core.app.ActivityCompat
 import androidx.core.app.NotificationManagerCompat
-import androidx.core.content.ContextCompat.startActivity
 import androidx.core.view.WindowCompat
 import androidx.documentfile.provider.DocumentFile
 import io.freetubeapp.freetube.MainActivity
 import io.freetubeapp.freetube.MediaControlsReceiver
 import io.freetubeapp.freetube.R
-import io.freetubeapp.freetube.WriteMode
-import io.freetubeapp.freetube.readFile
-import io.freetubeapp.freetube.readText
-import io.freetubeapp.freetube.writeFile
-import io.freetubeapp.freetube.writeText
+import io.freetubeapp.freetube.helpers.WriteMode
+import io.freetubeapp.freetube.helpers.readFile
+import io.freetubeapp.freetube.helpers.readText
+import io.freetubeapp.freetube.helpers.writeFile
+import io.freetubeapp.freetube.helpers.writeText
 import org.json.JSONObject
 import java.io.File
-import java.io.FileInputStream
 import java.net.URL
 import java.util.UUID.*
